Stop Bazel from generating an error when building under root
Unfortunately, python rules in Bazel, for reasons that are beyond the scope
of this comment, added a check that fails builds when it detects that Bazel
runs under root superuser.

It's not obvious how to prevent Dalec from using root super user (we tried
su command, but it wasn't successful), so instead we patch Envoy to
forecefully disable the super user check in python rules.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>

Signed-off-by: Gustavo <grnmeira@gmail.com>